# docs/openapi.yaml
openapi: 3.1.0
info:
  title: x4-api
  version: 0.0.0
paths:
...
  /api/v1/map/clusters:
    get:
      tags:
      - map
      summary: List Clusters
      operationId: list_clusters_api_v1_map_clusters_get
      parameters:
      - name: owner_faction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClusterSummary'
                title: Response List Clusters Api V1 Map Clusters Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/map/sectors:
    get:
      tags:
      - map
      summary: List Sectors
      operationId: list_sectors_api_v1_map_sectors_get
      parameters:
      - name: cluster_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
      - name: owner_faction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SectorSummary'
                title: Response List Sectors Api V1 Map Sectors Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/sectors/{sector_id}:
    get:
      tags:
      - map
      summary: Get Sector
      operationId: get_sector_api_v1_map_sectors__sector_id__get
      parameters:
      - name: sector_id
        in: path
        required: true
        schema:
          type: string
          title: Sector Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectorSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/player/reputation:
    get:
      tags:
      - player
      summary: Player Reputation
      description: The player's current standing with every faction (best first),
        with gamestart drift.
      operationId: player_reputation_api_v1_player_reputation_get
      parameters:
      - name: include_hidden
        in: query
        required: false
        schema:
          type: boolean
          description: Include utility/hidden factions.
          default: false
          title: Include Hidden
        description: Include utility/hidden factions.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlayerRelation'
                title: Response Player Reputation Api V1 Player Reputation Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ClusterSummary:
      properties:
        cluster_id:
          type: string
          title: Cluster Id
        macro_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Macro Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
        dlc:
          anyOf:
          - type: string
          - type: 'null'
          title: Dlc
        environment:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment
        sun_class:
          anyOf:
          - type: string
          - type: 'null'
          title: Sun Class
        population_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Population Id
        max_population:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Population
        x:
          anyOf:
          - type: number
          - type: 'null'
          title: X
        y:
          anyOf:
          - type: number
          - type: 'null'
          title: Y
        z:
          anyOf:
          - type: number
          - type: 'null'
          title: Z
        qx:
          anyOf:
          - type: number
          - type: 'null'
          title: Qx
        qy:
          anyOf:
          - type: number
          - type: 'null'
          title: Qy
        qz:
          anyOf:
          - type: number
          - type: 'null'
          title: Qz
        qw:
          anyOf:
          - type: number
          - type: 'null'
          title: Qw
      additionalProperties: false
      type: object
      required:
      - cluster_id
      title: ClusterSummary
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PlayerRelation:
      properties:
        faction_id:
          type: string
          title: Faction Id
        faction_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Faction Name
        color_hex:
          anyOf:
          - type: string
          - type: 'null'
          title: Color Hex
        relation:
          type: number
          title: Relation
        initial_relation:
          anyOf:
          - type: number
          - type: 'null'
          title: Initial Relation
      additionalProperties: false
      type: object
      required:
      - faction_id
      - faction_name
      - color_hex
      - relation
      - initial_relation
      title: PlayerRelation
...
    SectorSummary:
      properties:
        sector_id:
          type: string
          title: Sector Id
        cluster_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
        macro_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Macro Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
        dlc:
          anyOf:
          - type: string
          - type: 'null'
          title: Dlc
        sunlight:
          anyOf:
          - type: number
          - type: 'null'
          title: Sunlight
        economy:
          anyOf:
          - type: number
          - type: 'null'
          title: Economy
        security:
          anyOf:
          - type: number
          - type: 'null'
          title: Security
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
        access_licence:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Licence
        x:
          anyOf:
          - type: number
          - type: 'null'
          title: X
        y:
          anyOf:
          - type: number
          - type: 'null'
          title: Y
        z:
          anyOf:
          - type: number
          - type: 'null'
          title: Z
        qx:
          anyOf:
          - type: number
          - type: 'null'
          title: Qx
        qy:
          anyOf:
          - type: number
          - type: 'null'
          title: Qy
        qz:
          anyOf:
          - type: number
          - type: 'null'
          title: Qz
        qw:
          anyOf:
          - type: number
          - type: 'null'
          title: Qw
        known_to_player:
          type: boolean
          title: Known To Player
          default: false
      additionalProperties: false
      type: object
      required:
      - sector_id
      - cluster_id
      - owner_faction
      title: SectorSummary
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...